factor out C++ BoardClassBoggler
